Moving the residual path to the components, so that we can use the same definitions for a fused implementation through Triton -and components are enough to implement a Transformer-
