+++ xformers/components/__init__.py
@@ -13,6 +13,7 @@ from xformers.utils import import_all_modules
 from .activations import Activation, build_activation  # noqa
 from .attention import Attention, build_attention  # noqa
 from .multi_head_dispatch import MultiHeadDispatch, MultiHeadDispatchConfig  # noqa
+from .residual import LayerNormStyle, PostNorm, PreNorm, Residual  # noqa
 
 __all__ = ["MultiHeadDispatch", "Activation"]
